requirements.txt
# os2-server dependencies
python-telegram-bot>=21.0
pyyaml>=6.0
